+++ custom/plugins/SwagCmsExtensions/src/Resources/app/administration/src/extension/module/sw-cms/page/sw-cms-detail/index.js
@@ -13,7 +13,9 @@ Component.override('sw-cms-detail', {
         loadPageCriteria() {
             const criteria = this.$super('loadPageCriteria');
 
-            criteria.getAssociation('sections')
+            criteria
+                .addAssociation('swagCmsExtensionsScrollNavigationPageSettings')
+                .getAssociation('sections')
                 .addAssociation('swagCmsExtensionsScrollNavigation')
                 .getAssociation('blocks')
                 .addAssociation('swagCmsExtensionsQuickview');